+++ tests/client_connect_to_socket.rs
@@ -8,10 +8,7 @@ use std::os::unix::io::IntoRawFd;
 
 fn main() {
     let mut server = TestServer::new();
-    let loop_token = server.event_loop.token();
-    server
-        .display
-        .create_global::<ServerOutput, _>(&loop_token, 2, |_, _| {});
+    server.display.create_global::<ServerOutput, _>(2, |_, _| {});
 
     let (s1, s2) = ::std::os::unix::net::UnixStream::pair().unwrap();