+++ tests/helpers/mod.rs
@@ -11,16 +11,18 @@ use std::ffi::{OsStr, OsString};
 use std::io;
 use std::os::unix::io::RawFd;
 use std::rc::Rc;
+use std::time::Duration;
 
 pub struct TestServer {
     pub display: self::ways::Display,
-    pub event_loop: self::ways::EventLoop,
+    pub event_loop: self::ways::calloop::EventLoop<()>,
     pub socket_name: OsString,
 }
 
 impl TestServer {
     pub fn new() -> TestServer {
-        let (mut display, event_loop) = self::ways::Display::new();
+        let event_loop = self::ways::calloop::EventLoop::<()>::new().unwrap();
+        let mut display = self::ways::Display::new(event_loop.handle());
         let socket_name = display
             .add_socket_auto()
             .expect("Failed to create a server socket.");
@@ -33,10 +35,14 @@ impl TestServer {
     }
 
     pub fn answer(&mut self) {
-        self.event_loop.dispatch(Some(10)).unwrap();
+        self.event_loop
+            .dispatch(Some(Duration::from_millis(10)), &mut ())
+            .unwrap();
         self.display.flush_clients();
         // TODO: find out why native_lib requires two dispatches
-        self.event_loop.dispatch(Some(10)).unwrap();
+        self.event_loop
+            .dispatch(Some(Duration::from_millis(10)), &mut ())
+            .unwrap();
         self.display.flush_clients();
     }
 }