+++ wayland-server/src/native_lib/display.rs
@@ -9,19 +9,26 @@ use std::rc::Rc;
 
 use wayland_sys::server::*;
 
+use calloop::generic::Generic;
+use calloop::{LoopHandle, Source};
+
+use Fd;
+
 use super::globals::GlobalData;
-use super::{ClientInner, EventLoopInner, GlobalInner};
+use super::{ClientInner, GlobalInner};
 
 use display::get_runtime_dir;
 use {Interface, NewResource};
 
 pub(crate) struct DisplayInner {
     pub(crate) ptr: *mut wl_display,
+    source: Option<Source<Generic<Fd>>>,
 }
 
 impl Drop for DisplayInner {
     fn drop(&mut self) {
         {
+            self.source.take().map(|s| s.remove());
             unsafe {
                 ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_destroy, self.ptr);
             }
@@ -30,10 +37,9 @@ impl Drop for DisplayInner {
 }
 
 impl DisplayInner {
-    pub(crate) fn new() -> (Rc<RefCell<DisplayInner>>, EventLoopInner) {
+    pub(crate) fn new<Data: 'static>(handle: LoopHandle<Data>) -> Rc<RefCell<DisplayInner>> {
         unsafe {
             let ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_create,);
-            let display = Rc::new(RefCell::new(DisplayInner { ptr: ptr }));
             // setup the client_created listener
             let listener = signal::rust_listener_create(client_created);
             ffi_dispatch!(
@@ -51,10 +57,22 @@ impl DisplayInner {
                 ::std::ptr::null_mut()
             );
 
-            // extract the event loop
-            let evq_ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_get_event_loop, ptr);
-            let evq = EventLoopInner::display_new(display.clone(), evq_ptr);
-            (display, evq)
+            let evl_ptr = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_display_get_event_loop, ptr);
+            let evl_fd = ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_event_loop_get_fd, evl_ptr);
+
+            let mut evtsrc = Generic::new(Fd(evl_fd));
+            evtsrc.set_interest(::mio::Ready::readable());
+            evtsrc.set_pollopts(::mio::PollOpt::edge());
+
+            let source = Some(
+                handle
+                    .insert_source(evtsrc, move |_, _| {
+                        ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_event_loop_dispatch, evl_ptr, 0);
+                    })
+                    .unwrap(),
+            );
+
+            Rc::new(RefCell::new(DisplayInner { ptr, source }))
         }
     }
 
@@ -64,7 +82,6 @@ impl DisplayInner {
 
     pub(crate) fn create_global<I: Interface, F1, F2>(
         &mut self,
-        evl: &EventLoopInner,
         version: u32,
         implementation: F1,
         filter: Option<F2>,