+++ wayland-server/src/rust_imp/display.rs
@@ -2,21 +2,20 @@ use std::cell::RefCell;
 use std::env;
 use std::ffi::{OsStr, OsString};
 use std::io;
-use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd, RawFd};
+use std::os::unix::io::{FromRawFd, IntoRawFd, RawFd};
 use std::os::unix::net::UnixListener;
 use std::path::Path;
 use std::rc::Rc;
 
-use mio::Ready;
+use calloop::{LoopHandle, Source};
 
 use display::get_runtime_dir;
-use sources::{FdEvent, FdInterest};
 use {Interface, NewResource};
 
 use super::clients::ClientManager;
-use super::event_loop::SourcesPoll;
+use super::event_loop_glue::{WSLoopHandle, WaylandListener};
 use super::globals::GlobalManager;
-use super::{ClientInner, EventLoopInner, GlobalInner, SourceInner};
+use super::{ClientInner, GlobalInner};
 
 pub(crate) const DISPLAY_ERROR_INVALID_OBJECT: u32 = 0;
 pub(crate) const DISPLAY_ERROR_INVALID_METHOD: u32 = 1;
@@ -24,36 +23,31 @@ pub(crate) const DISPLAY_ERROR_INVALID_METHOD: u32 = 1;
 pub(crate) const DISPLAY_ERROR_NO_MEMORY: u32 = 2;
 
 pub(crate) struct DisplayInner {
-    sources_poll: SourcesPoll,
-    clients_mgr: Rc<RefCell<ClientManager>>,
+    loophandle: Box<WSLoopHandle>,
+    pub(crate) clients_mgr: Rc<RefCell<ClientManager>>,
     global_mgr: Rc<RefCell<GlobalManager>>,
-    listeners: Vec<SourceInner<FdEvent>>,
+    listeners: Vec<Source<WaylandListener>>,
 }
 
 impl DisplayInner {
-    pub(crate) fn new() -> (Rc<RefCell<DisplayInner>>, EventLoopInner) {
-        let mut evl = EventLoopInner::new();
-
+    pub(crate) fn new<Data: 'static>(handle: LoopHandle<Data>) -> Rc<RefCell<DisplayInner>> {
         let global_mgr = Rc::new(RefCell::new(GlobalManager::new()));
 
         let display = Rc::new(RefCell::new(DisplayInner {
-            sources_poll: evl.get_poll(),
+            loophandle: Box::new(handle.clone()),
             clients_mgr: Rc::new(RefCell::new(ClientManager::new(
-                evl.get_poll(),
+                Box::new(handle.clone()),
                 global_mgr.clone(),
             ))),
             global_mgr,
             listeners: Vec::new(),
         }));
 
-        evl.display = Some(display.clone());
-
-        (display, evl)
+        display
     }
 
     pub(crate) fn create_global<I: Interface, F1, F2>(
         &mut self,
-        evl: &EventLoopInner,
         version: u32,
         implementation: F1,
         filter: Option<F2>,
@@ -64,7 +58,7 @@ impl DisplayInner {
     {
         self.global_mgr
             .borrow_mut()
-            .add_global(evl, version, implementation, filter)
+            .add_global(version, implementation, filter)
     }
 
     pub(crate) fn flush_clients(&mut self) {
@@ -72,23 +66,15 @@ impl DisplayInner {
     }
 
     fn add_unix_listener(&mut self, listener: UnixListener) -> io::Result<()> {
-        let fd = listener.as_raw_fd();
-
         listener.set_nonblocking(true)?;
 
-        let mut implem = ListenerImplementation {
-            listener,
-            client_mgr: self.clients_mgr.clone(),
-        };
+        let client_mgr = self.clients_mgr.clone();
 
-        let source = self.sources_poll.insert_source(
-            fd,
-            Ready::readable(),
-            move |evt| implem.receive(evt),
-            FdEvent::Ready {
-                fd,
-                mask: FdInterest::READ,
-            },
+        let source = self.loophandle.add_listener(
+            WaylandListener::new(listener),
+            Box::new(move |stream| unsafe {
+                client_mgr.borrow_mut().init_client(stream.into_raw_fd());
+            }),
         )?;
 
         self.listeners.push(source);
@@ -143,69 +129,10 @@ impl DisplayInner {
     }
 }
 
-struct ListenerImplementation {
-    listener: UnixListener,
-    client_mgr: Rc<RefCell<ClientManager>>,
-}
-
-impl ListenerImplementation {
-    fn eprint_error(&self, verb: &str, error: io::Error) {
-        if let Ok(addr) = self.listener.local_addr() {
-            if let Some(path) = addr.as_pathname() {
-                eprintln!(
-                    "[wayland-server] Error {} listening socket {} : {}",
-                    verb,
-                    path.display(),
-                    error
-                );
-                return;
-            }
-        }
-        eprintln!(
-            "[wayland-server] Error {} listening socket <unnamed> : {}",
-            verb, error
-        );
-    }
-
-    fn receive(&mut self, event: FdEvent) {
-        match event {
-            FdEvent::Ready { .. } => {
-                // one (or more) clients connected to the socket
-                loop {
-                    match self.listener.accept() {
-                        Ok((stream, _)) => unsafe {
-                            self.client_mgr.borrow_mut().init_client(stream.into_raw_fd());
-                        },
-                        Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
-                            // we have exhausted all the pending connections
-                            break;
-                        }
-                        Err(e) => {
-                            // this is a legitimate error
-                            self.eprint_error("accepting connection on", e);
-                        }
-                    }
-                }
-            }
-            FdEvent::Error { error, .. } => self.eprint_error("polling", error),
-        }
-    }
-}
-
 impl Drop for DisplayInner {
     fn drop(&mut self) {
         for l in self.listeners.drain(..) {
             l.remove();
-        }
-    }
-}
-
-impl Drop for ListenerImplementation {
-    fn drop(&mut self) {
-        if let Ok(socketaddr) = self.listener.local_addr() {
-            if let Some(path) = socketaddr.as_pathname() {
-                let _ = ::std::fs::remove_file(path);
-            }
         }
     }
 }