+++ wayland-server/src/rust_imp/mod.rs
@@ -11,13 +11,12 @@ use {Interface, NewResource, Resource};
 
 mod clients;
 mod display;
-mod event_loop;
+mod event_loop_glue;
 mod globals;
 mod resources;
 
 pub(crate) use self::clients::ClientInner;
 pub(crate) use self::display::DisplayInner;
-pub(crate) use self::event_loop::{EventLoopInner, IdleSourceInner, SourceInner};
 pub(crate) use self::globals::GlobalInner;
 pub(crate) use self::resources::{NewResourceInner, ResourceInner};